Move unit testing to pytest
Using the Pytest integration to Marimo, we can still author unit tests
in Marimo notebooks. The support of these for the pytest-asyncio plug-in
is very bad, however, so all tests encapsulate async computations. This
is enabled through the implementation of a bi-color iterator, which can
handle either the native async iteration of an applied chain, or
dissimulate the instantiation of an event loop in an ad hoc thread.